Module aws_sdk_licensemanager::error

Modules

access_denied_exception

See AccessDeniedException

authorization_exception

See AuthorizationException

conflict_exception

See ConflictException

entitlement_not_allowed_exception

See EntitlementNotAllowedException

failed_dependency_exception

See FailedDependencyException

filter_limit_exceeded_exception

See FilterLimitExceededException

invalid_parameter_value_exception

See InvalidParameterValueException

invalid_resource_state_exception

See InvalidResourceStateException

license_usage_exception

See LicenseUsageException

no_entitlements_allowed_exception

See NoEntitlementsAllowedException

rate_limit_exceeded_exception

See RateLimitExceededException

redirect_exception

See RedirectException

resource_limit_exceeded_exception

See ResourceLimitExceededException

resource_not_found_exception

See ResourceNotFoundException

server_internal_exception

See ServerInternalException

unsupported_digital_signature_method_exception

See UnsupportedDigitalSignatureMethodException

validation_exception

See ValidationException

Structs

AcceptGrantError

AccessDeniedException

Access to resource denied.

AuthorizationException

The AWS user account does not have permission to perform the action. Check the IAM

CheckInLicenseError

CheckoutBorrowLicenseError

CheckoutLicenseError

ConflictException

There was a conflict processing the request. Try your request again.

CreateGrantError

CreateGrantVersionError

CreateLicenseConfigurationError

CreateLicenseError

CreateLicenseManagerReportGeneratorError

CreateLicenseVersionError

CreateTokenError

DeleteGrantError

DeleteLicenseConfigurationError

DeleteLicenseError

DeleteLicenseManagerReportGeneratorError

DeleteTokenError

EntitlementNotAllowedException

The entitlement is not allowed.

ExtendLicenseConsumptionError

FailedDependencyException

A dependency required to run the API is missing.

FilterLimitExceededException

The request uses too many filters or too many filter values.

GetAccessTokenError

GetGrantError

GetLicenseConfigurationError

GetLicenseError

GetLicenseManagerReportGeneratorError

GetLicenseUsageError

GetServiceSettingsError

InvalidParameterValueException

One or more parameter values are not valid.

InvalidResourceStateException

License Manager cannot allocate a license to a resource because of its state.

LicenseUsageException

You do not have enough licenses available to support a new resource launch.

ListAssociationsForLicenseConfigurationError

ListDistributedGrantsError

ListFailuresForLicenseConfigurationOperationsError

ListLicenseConfigurationsError

ListLicenseManagerReportGeneratorsError

ListLicenseSpecificationsForResourceError

ListLicenseVersionsError

ListLicensesError

ListReceivedGrantsError

ListReceivedLicensesError

ListResourceInventoryError

ListTagsForResourceError

ListTokensError

ListUsageForLicenseConfigurationError

NoEntitlementsAllowedException

There are no entitlements found for this license, or the entitlement maximum count is reached.

RateLimitExceededException

Too many requests have been submitted. Try again after a brief wait.

RedirectException

This is not the correct Region for the resource. Try again.

RejectGrantError

ResourceLimitExceededException

Your resource limits have been exceeded.

ResourceNotFoundException

The resource cannot be found.

ServerInternalException

The server experienced an internal error. Try again.

TagResourceError

UnsupportedDigitalSignatureMethodException

The digital signature method is unsupported. Try your request again.

UntagResourceError

UpdateLicenseConfigurationError

UpdateLicenseManagerReportGeneratorError

UpdateLicenseSpecificationsForResourceError

UpdateServiceSettingsError

ValidationException

The provided input is not valid. Try your request again.

Enums

AcceptGrantErrorKind

CheckInLicenseErrorKind

CheckoutBorrowLicenseErrorKind

CheckoutLicenseErrorKind

CreateGrantErrorKind

CreateGrantVersionErrorKind

CreateLicenseConfigurationErrorKind

CreateLicenseErrorKind

CreateLicenseManagerReportGeneratorErrorKind

CreateLicenseVersionErrorKind

CreateTokenErrorKind

DeleteGrantErrorKind

DeleteLicenseConfigurationErrorKind

DeleteLicenseErrorKind

DeleteLicenseManagerReportGeneratorErrorKind

DeleteTokenErrorKind

ExtendLicenseConsumptionErrorKind

GetAccessTokenErrorKind

GetGrantErrorKind

GetLicenseConfigurationErrorKind

GetLicenseErrorKind

GetLicenseManagerReportGeneratorErrorKind

GetLicenseUsageErrorKind

GetServiceSettingsErrorKind

ListAssociationsForLicenseConfigurationErrorKind

ListDistributedGrantsErrorKind

ListFailuresForLicenseConfigurationOperationsErrorKind

ListLicenseConfigurationsErrorKind

ListLicenseManagerReportGeneratorsErrorKind

ListLicenseSpecificationsForResourceErrorKind

ListLicenseVersionsErrorKind

ListLicensesErrorKind

ListReceivedGrantsErrorKind

ListReceivedLicensesErrorKind

ListResourceInventoryErrorKind

ListTagsForResourceErrorKind

ListTokensErrorKind

ListUsageForLicenseConfigurationErrorKind

RejectGrantErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateLicenseConfigurationErrorKind

UpdateLicenseManagerReportGeneratorErrorKind

UpdateLicenseSpecificationsForResourceErrorKind

UpdateServiceSettingsErrorKind