Module aws_sdk_licensemanager::model

Modules

automated_discovery_information

See AutomatedDiscoveryInformation

borrow_configuration

See BorrowConfiguration

consumed_license_summary

See ConsumedLicenseSummary

consumption_configuration

See ConsumptionConfiguration

datetime_range

See DatetimeRange

entitlement

See Entitlement

entitlement_data

See EntitlementData

entitlement_usage

See EntitlementUsage

filter

See Filter

grant

See Grant

granted_license

See GrantedLicense

inventory_filter

See InventoryFilter

issuer

See Issuer

issuer_details

See IssuerDetails

license

See License

license_configuration

See LicenseConfiguration

license_configuration_association

See LicenseConfigurationAssociation

license_configuration_usage

See LicenseConfigurationUsage

license_operation_failure

See LicenseOperationFailure

license_specification

See LicenseSpecification

license_usage

See LicenseUsage

managed_resource_summary

See ManagedResourceSummary

metadata

See Metadata

organization_configuration

See OrganizationConfiguration

product_information

See ProductInformation

product_information_filter

See ProductInformationFilter

provisional_configuration

See ProvisionalConfiguration

received_metadata

See ReceivedMetadata

report_context

See ReportContext

report_frequency

See ReportFrequency

report_generator

See ReportGenerator

resource_inventory

See ResourceInventory

s3_location

See S3Location

tag

See Tag

token_data

See TokenData

Structs

AutomatedDiscoveryInformation

Describes automated discovery.

BorrowConfiguration

Details about a borrow configuration.

ConsumedLicenseSummary

Details about license consumption.

ConsumptionConfiguration

Details about a consumption configuration.

DatetimeRange

Describes a time range, in ISO8601-UTC format.

Entitlement

Describes a resource entitled for use with a license.

EntitlementData

Data associated with an entitlement resource.

EntitlementUsage

Usage associated with an entitlement resource.

Filter

A filter name and value pair that is used to return more specific results from a

Grant

Describes a grant.

GrantedLicense

Describes a license that is granted to a grantee.

InventoryFilter

An inventory filter.

Issuer

Details about the issuer of a license.

IssuerDetails

Details associated with the issuer of a license.

License

Software license that is managed in AWS License Manager.

LicenseConfiguration

A license configuration is an abstraction of a customer license agreement that can be

LicenseConfigurationAssociation

Describes an association with a license configuration.

LicenseConfigurationUsage

Details about the usage of a resource associated with a license configuration.

LicenseOperationFailure

Describes the failure of a license operation.

LicenseSpecification

Details for associating a license configuration with a resource.

LicenseUsage

Describes the entitlement usage associated with a license.

ManagedResourceSummary

Summary information about a managed resource.

Metadata

Describes key/value pairs.

OrganizationConfiguration

Configuration information for AWS Organizations.

ProductInformation

Describes product information for a license configuration.

ProductInformationFilter

Describes product information filters.

ProvisionalConfiguration

Details about a provisional configuration.

ReceivedMetadata

Metadata associated with received licenses and grants.

ReportContext

Details of the license configuration that this generator reports on.

ReportFrequency

Details on how frequently reports are generated.

ReportGenerator

Describe the details of a report generator.

ResourceInventory

Details about a resource.

S3Location

Details of the S3 bucket that report generator reports are published to.

Tag

Details about a tag for a license configuration.

TokenData

Describes a token.

Enums

AllowedOperation

CheckoutType

DigitalSignatureMethod

EntitlementDataUnit

EntitlementUnit

GrantStatus

InventoryFilterCondition

LicenseConfigurationStatus

LicenseCountingType

LicenseDeletionStatus

LicenseStatus

ReceivedStatus

RenewType

ReportFrequencyType

ReportType

ResourceType

TokenType