Module aws_sdk_licensemanager::operation

Structs

AcceptGrant

Accepts the specified grant.

CheckInLicense

Checks in the specified license. Check in a license when it is no longer in use.

CheckoutBorrowLicense

Checks out the specified license for offline use.

CheckoutLicense

Checks out the specified license.

CreateGrant

Creates a grant for the specified license. A grant shares the use of license entitlements with specific AWS accounts.

CreateGrantVersion

Creates a new version of the specified grant.

CreateLicense

Creates a license.

CreateLicenseConfiguration

Creates a license configuration.

CreateLicenseManagerReportGenerator

Creates a new report generator.

CreateLicenseVersion

Creates a new version of the specified license.

CreateToken

Creates a long-lived token.

DeleteGrant

Deletes the specified grant.

DeleteLicense

Deletes the specified license.

DeleteLicenseConfiguration

Deletes the specified license configuration.

DeleteLicenseManagerReportGenerator

Delete an existing report generator.

DeleteToken

Deletes the specified token. Must be called in the license home Region.

ExtendLicenseConsumption

Extends the expiration date for license consumption.

GetAccessToken

Gets a temporary access token to use with AssumeRoleWithWebIdentity. Access tokens

GetGrant

Gets detailed information about the specified grant.

GetLicense

Gets detailed information about the specified license.

GetLicenseConfiguration

Gets detailed information about the specified license configuration.

GetLicenseManagerReportGenerator

Gets information on the specified report generator.

GetLicenseUsage

Gets detailed information about the usage of the specified license.

GetServiceSettings

Gets the License Manager settings for the current Region.

ListAssociationsForLicenseConfiguration

Lists the resource associations for the specified license configuration.

ListDistributedGrants

Lists the grants distributed for the specified license.

ListFailuresForLicenseConfigurationOperations

Lists the license configuration operations that failed.

ListLicenseConfigurations

Lists the license configurations for your account.

ListLicenseManagerReportGenerators

Lists the report generators for your account.

ListLicenseSpecificationsForResource

Describes the license configurations for the specified resource.

ListLicenseVersions

Lists all versions of the specified license.

ListLicenses

Lists the licenses for your account.

ListReceivedGrants

Lists grants that are received but not accepted.

ListReceivedLicenses

Lists received licenses.

ListResourceInventory

Lists resources managed using Systems Manager inventory.

ListTagsForResource

Lists the tags for the specified license configuration.

ListTokens

Lists your tokens.

ListUsageForLicenseConfiguration

Lists all license usage records for a license configuration, displaying license

RejectGrant

Rejects the specified grant.

TagResource

Adds the specified tags to the specified license configuration.

UntagResource

Removes the specified tags from the specified license configuration.

UpdateLicenseConfiguration

Modifies the attributes of an existing license configuration.

UpdateLicenseManagerReportGenerator

Updates a report generator.

UpdateLicenseSpecificationsForResource

Adds or removes the specified license configurations for the specified AWS resource.

UpdateServiceSettings

Updates License Manager settings for the current Region.